Move `SpringModule` from `utbot-framework` to `utbot-spring-framework`
